// apps/api/src/domain/enums/estado-processo.enum.ts
export enum EstadoProcesso {
  RASCUNHO = "rascunho",
  EM_INSTRUCAO = "em_instrucao",
  REMETIDO_PROCURADORIA = "remetido_procuradoria",
  EM_PROCURADORIA = "em_procuradoria",
  DESPACHADO = "despachado",
  CUMPRIDO = "cumprido",
  ACUSADO = "acusado",
  ARQUIVADO = "arquivado",
  SUSPENSO = "suspenso",
  EXTINTO = "extinto"
}